Eliminate random modules. Refs #157.
Copilot-core currently includes facilities to generate random streams.
However, the new testing system being created by @fdedden, based on
QuickCheck, uses standard quickcheck generators to create random
streams, and does not use the implementation currently in copilot-core.
While the current implementation may be useful in the future, it does
not currently make sense that this be a part of the copilot
distribution.

This change removes all (old) Random-related modules from the tree.