Added superfluous comma in CSV output of driver. Refs #184.
The interpreter adds a comma after the triggername (in CSV output), even if the
trigger does not have any arguments. In CSV this implies that there is an empty
argument, it shouldn't happen. The interpreter will be fixed later on, for now
we just add a comma to the output of the driver as well.