Added the long printf format for 64 bit numbers. Refs #177.
Ideally this should all be platform independent (so we can test other
architectures as well). Printing numbers using printf in C is however not
easily translated to other architectures, as the size of a long, for example,
might be different on other platforms. The C99 standard helps in using with the
macro's defined in <inttypes.h>, these should be used as soon as our C99
implementation supports macros.