Simplified arguments to mkidents. Refs #176.
The usage of a tuple was initially a good idea, as the namedargs lists is
tupled as well. However for recursive types, we would like to call mkidents
recursively. Having a UExpr as an argument to mkidents doesn't work in this
case: there is no UExpr for recursive types. Replacing this by the Type
containted in the UExpr fixes this. Now there is no reason to use a tuple
anymore as well.