Document the package and all modules consistently. Refs #67.
A lot of the code in the copilot package is undocumented. This makes it
hard for others to navigate the API, and even for ourselves to maintain
it.

This patch documents all modules in this package consistently, and the
package declaration itself. Note that this patch does not address any
other aspects of code cleaning or formatting; those can be addressed by a
separate issue.