Remove version bounds for copilot package in examples. Refs #86.
The examples included as executables in the cabal file depend on
specific versions of copilot. This declaration is not necessary, since
the examples are in the same cabal file as the library. Cabal reports
these version bounds as extraneous.

This change removes all version bounds for all copilot packages in all
the examples.