Remove unused function. Refs #94.
The function copilot-theorem:Copilot.Theorem.Misc.Error.notHandled is
not being used by any other code in the Copilot project.

Although the function is exported by the module it is defined in, the
module itself is not being exported by the library, so there is no need
for deprecation. The function notHandled can be removed right away.

This change removes the function from the export list and the module.