Adds support for what4 1.1. Refs #216.
This commit makes the changes required to the what4 translation for what4-1.1
compatibility.

There were two significant changes that affected Copilot.Theorem.What4. The
first was the type of ExprSymFn changed trivially; it no longer requires (Expr
t) as an argument. The second was that concrete floating point operations are
now implemented, and the libBF library is used for that, so several types
changed from raw bitvectors to libBF floats.