Remove second element of pair FunEnv. Refs #170.
The second element of the pair that defines the type synonym
Copilot.Compile.C99.Util.FunEnv is not being used anywhere in the code.

The FunEnv is used as the state in a state monad, but the only
invocation to change the state in the monad puts an empty list in that
part of the tuple. That part is also ignored in the return value when
the state computation is executed.

This commit changes the definition and all invocations to match the new
type of the function and the state in the aforementioned monad.