Deprecate internal modules in copilot-c99. Refs #237.
The copilot-c99 backend exposes internal modules, types and operations.
In principle, only the Copilot.Compile.C99 module needs to be exposed.

In preparation for the removal of internal definitions from the exposed
API, this change deprecates all internal modules in copilot-c99, so that
any direct imports trigger a warning for the user.

This copilot-c99.cabal file is adapted to disable warnings during
compilation, so that the warnings do not lead users to believe that
something is internally wrong in the library.