Rename modules to avoid transitive deprecation warnings. Refs #242.
The implementation of the fix for #237 introduced an issue: users are
now getting undesired deprecation warnings.

The problem stems from a misunderstanding of how the deprecation option
would be handled by GHC. In short: GHC makes it transitive, but my
(incorrect) expectation was that it wouldn't be. Further details can be
found at:
https://mail.haskell.org/pipermail/haskell-cafe/2021-August/134333.html

This patch addresses that issue by moving all definitions re-exported by
the module Copilot.Compile.C99 into submodules, which are not
re-exported directly.