Re-organize project in folders. Refs #248.
This commit moves the main copilot library to a subfolder, and moves the
dependencies in lib/ to the top level, each to its own directory.

The README file, how inside the copilot folder is linked to from the top
level using a symbolic link, so that github can find it and shows it
to repo visitors up-front.

The cabal.project file is updated to reflect the new repo layout.