Update Travis CI config to work in new repo layout. Refs #248.
The travis configuration file assumes that the Copilot library
is located at the top level, so Cabal implicitly knows what
the main library is. That assumption is no longer true after
re-structuring the repo.

This commit updates the travis file to the new repo layout, by
explicitly indicating which component must be built.