Fix broken links to LICENSES in READMEs. Refs #251.
The links to the licenses in the READMEs (including the top level
README) have been either outdated or broken as a result of several repo
reorganization processes.

This commit updates all links to LICENSE files in the READMEs.

Each README points to the corresponding package's own LICENSE file, as
opposed to pointing to the main Copilot LICENSE.

[ci skip]