Fix broken links to Travis in READMEs. Refs #251.
The links to the Travis badge and/or Travis CI location for Copilot in
the READMEs (including the top level README) have been either outdated
or broken as a result of several repo reorganization processes.

This commit updates all links to Travis resources in the READMEs. All
READMEs point to the same locations on Travis, since compilation and CI
applies and is carried out on Travis CI for the whole project as an
indivisible unit.

[ci skip]