Update links in Haskell haddock. Refs #251.
Some links to different repo locations in the Haddock comments that
accompany the code are outdated, and point to locations that have been
either outdated or broken as a result of several repo reorganization
processes.

This commit updates links in the code comments. Where a new version of
the file being mentioned did not exist in the current development
branch, the link has been adjusted to point to a working location using
the last known tag containing that file.

[ci skip]