Lowercase repo name for consistency. Refs #251.
The repo name is sometimes refered as Copilot and other times as copilot
(in lowercase).

Using different cases for the same resource or element makes it harder
to find expressions in the repo using standard tools.

This commit simply updates the repo URL in the main cabal file and in
the README to consistently use lowercase for the repo name.

[ci skip]