Add a `cSettingOutputDirectory` setting for codegen. Refs #255.
The code generator assumes that the output directory for the generated C
files must be the present directory. Although the filename prefix can be
customized, one cannot add a directory separator to it, since it is used
also in the CPP includes in the generated C code.

This commit introduces a new field to the CSettings type that allows
users to define an output destination directory for generated files. It
will be created if necessary.

Co-authored-by: Rob Dockins <rdockins@galois.com>
Co-authored-by: Ivan Perez <ivan.perezdominguez@nasa.gov>