Unit testing of copilot-core:Copilot.Core.Interpret.Eval. Refs #256.
This commit introduces tests for the Copilot.Core.Interpret.Eval module,
checking the evaluation of stream expressions up to a predefined length
of the trace.

The test introduces a notion of a semantics pair, which is a pair of an
object whose implementation we are checking and an object in the
denotational domain.

For the purposes of testing the evaluator, we consider that pure values
(elements carried by a stream) are identical in the source and target
domains of the semantics function; the meaning of an expression is an
infinite list of values, and the meaning of an operator is a pointwise
function on (list) values.