Deprecate type Tag. Refs #260.
The type synonym copilot-core:Copilot.Core.Expr.Tag is not being used
anywhere in the code of copilot or the examples. Searching for it in the
project shows only two occurrences of the same word beyond its own
definition and export in the module: one as part of a string, and the
other as part of a comment.

This commit deprecates the type so it can be safely removed in the
future, to minimize impact to users since it is being exposed.