Unit testing of copilot-core:Copilot.Core.Type. Refs #256.
This commit introduces tests for the Copilot.Core.Type module.

The module introduces Type (and Struct and Array) and SimpleType.

Struct and Array have auxiliary operations that can be tested. However,
building meaningful examples has proved hard due to difficulty
convincing GHC's type checker of the types of expressions. Therefore,
the interface for Type remains untested. We rely on other parts of the
tests in other modules to test the Type interface.

The API for SimpleType is small and, apart from its association with
Type, only provides an instance of equality.

This commit introduces a test of that instance constituting an actual
equality relation, for all types except arrays.