Introduce support for atan2. Refs #246.
Copilot currently support a limited number of operations. Although some
trigonometric functions are supported, atan2 is a common operation that
Copilot does not currently support.

This commit introduces atan2 as a new Op2, and exposes it all the way to
the user. Unlike operations defined in the Num and other type classes,
the type class in which atan2 is defined in Haskell includes operations
that do not make sense for streams (because the return type is a fixed
number, as opposed to a stream of numbers). This makes using the same
type class to overload the function potentially dangerous.

The commit also introduces unit tests for atan2 at core level.