Improve documentation of LTL module. Refs #131.
The LTL library is implemented using drops. Due to the way that Copilot
works, drops cannot always be applied to the result of a pointwise
transformation on streams, as Copilot has no way of determining if the
are enough samples to drop from. This results in a limited ability to
combine applications of temporal operators to build up complex formulas.

This commit improves the documentation of the LTL module to include the
aforementioned information. It will hopefully help users understand why
always and eventually may crash even when there are apparently enough
samples available for Copilot to simulate the predicate.

[ci skip]