Deprecate the export of Copilot.Core.Error. Refs #267.
This commit deprecates the functions in copilot-core:Copilot.Core.Error.
It introduces an internal replacement module that implements the same
features and subsitutes all uses of the aforementioned module to point
to the internal implementation, so that the one currently exported can
be DEPRECATED without giving rise to unintended compilation warnings.