Deprecate export of module copilot-core:Copilot.Core.Interpret.Render. Refs #268.
The module Copilot.Core.Interpret.Render defines two auxiliary functions
that are meant to be used via Copilot.Core.Interpret.

There is no reason to export that module, so it's export is exported. A
temporary module keeping the features implemented is included until the
module is hidden from the public interface.