Deprecate copilot-core:Copilot.Core.Type.Uninitialized.uninitialized. Refs #270.
The module copilot-core:Copilot.Core.Type.Uninitialized is only ever
used in one example, which currently does not compile due to needing Z3.

The current implementation is not complete either (#138), so its use by
users is limited.

As we do not have a use case that requires supporting this function,
it's best to remove it, and copy the code that needs it to the example
that actually needs this code.