Add a copilot-c99-oriented structs example. Refs #275.
This example is similar to the `what4-structs` example, except that this
example replaces the What4 properties with triggers, and instead of verifying
properties, this example compiles the specification to C code using
`copilot-c99`. Importantly, the generated C code will not compile without the
fix for #275, so this example also serves as a regression test of sorts.