copilot-c99: Use more specific operations for floats. Refs #263.
Previously, the C99 code generation would use the same operations for both
`double`s and `float`s, which would result in unnecessary numerical coercions.
This patch teaches the code generator to check the argument type to pick the
most suitable operation. For the details, see the Haddocks for the new
`transopFP` helper function. Almost all of the floating-point operations are
now implemented in terms of `transopFP`. Some exceptions are:

* `Abs`: `Abs` has a sensible interpretation for integral types (unlike the
  other floating-point operations), so we implement a special case for it.
* `Recip`: `Recip` is implemented in terms of `(./)`, which doesn't quite fit
  the pattern set in `transopFP`. Again, we implement a special case for it.