Remove deprecated function Copilot.Core.Type.Show.showWit. Refs #282.
The function Copilot.Core.Type.Show.showWit was marked as deprecated in
Copilot 3.4, so it is ready to be hidden from the interface (it is still
needed by a function that is exported).

As a consequence, the type ShowWit can now be deprecated, since there is
no (public) way to use it.