Make imports explicit. Reorganize imports. Refs #277.
Copilot currently fails to compile with newer versions of GHC because
there are new definitions in base that clash with definitions in
Copilot.

This commit re-organizes imports in the
copilot-core:Copilot.Core.Interpret.Eval module, and makes imports
explicit, to prevent future additions in base to clash with definitions
in Copilot.