Guard against compilation to C of specs without triggers. Refs #274.
Presently, copilot crashes when trying to compile an empty spec (one
without triggers) to C99. The error message emerges from a use of
fromList on an empty list in language-c99-simple, which does not help
Copilot users understand what failed, or how to fix it.

This commit checks for specs without triggers when trying to compile to
C99, and produces a more meaningful error message.