Mark packages as uncurated. Refs #288.
Packages should be published exactly as we have, without further
modifications, and nobody outside our project should make any changes
without our explicit authorization.

As of currently, the Hackage "policy" (whatever that means) implies that
Hackage maintainers should be able to correct issues with the cabal
files. This is against our own policies.

This commit sets the field `x-curation` to `uncurated` in the cabal
files of all libraries in the Copilot project, to indicate to Hackage
maintainers that our packages are to be left untouched.