Replace exposed modules with old internal implementations. Refs #289.
In a prior commit, internal implementations had been moved to a separate
module so that modules that were not going to be exposed directly could
be deprecated safely without triggering warnings when compiling
copilot-c99.

With the concealment of the library's implementation, those additional
internal modules can return to where they were. This commit renames the
modules to have their original names, adjusts imports, and the cabal
file so that the modules removed are no longer listed in the library.