Remove depracation notices. Refs #289.
Several internal modules had deprecation notices to warn users that they
were going to be removed.

This commit removes those notices, since the modules have now been
hidden and the only possible way those warnings could be triggered is
internally during compilation of this copilot-c99 library, which is not
desirable.