Test reification of boolean streams. Refs #271.
This commit implements tests to reify boolean streams, and compare them
against their ideal meaning. The meaning is given as a list of elements.

The implementation mirrors copilot-core's, except that there is now an
intermediate step of reification, which happens in the IO monad, and it
is not necessary to use typeOf to obtain the types of expressions.
Another difference wrt. the tests in copilot-core is that streams do not
implement a way to be shown.