Remove module Copilot.Core.Type.Uninitialized. Refs #302.
The function Copilot.Core.Type.Uninitialized.uninitialized was
deprecated in Copilot 3.6. The function was incomplete, and it was only
used in one example, so what was needed from it was moved directly
there.

As per our internal policy of waiting 3 versions from deprecation until
a public interface declaration can be removed, this type can now be
removed completely. Since it is the only definition in the module
Copilot.Core.Type.Uninitialized, the whole module can be removed.

The following demonstrates that the module is unused within Copilot:

```
$ grep -nHre 'Uninitialized' --include='*.hs' | grep -ve '^{-#' | grep -ve '--.*Uninitialized' | grep -ve '"[^"]*Uninitialized[^"]*"'
copilot-core/src/Copilot/Core/Type/Uninitialized.hs:10:module Copilot.Core.Type.Uninitialized
copilot-core/src/Copilot/Core.hs:33:  , module Copilot.Core.Type.Uninitialized
copilot-core/src/Copilot/Core.hs:45:import Copilot.Core.Type.Uninitialized
```

We use the module and not the function because a function with the same
name is used in an example.

The result obtained are three lines: the module, its re-import from
Copilot.Core, and it being exported. The function uninitialized is not
required by Copilot's requirements, and therefore should not be
published.

This commit removes the module, and removes its declaration as a
re-exported module both in Copilot.Core and directly in the cabal
package. Using the method above shows no results.