copilot-c99: Initialize arrays and structs without explcit casts. Refs #276.
Previously, C99 translation would produce initializers for array and
struct that were explicitly cast. Explicit casts of arrays and structs
are considered non-constant expressions in C, and thus not allowed as
the initialization value for a global variable. This causes the
generated code to produce warnings when compiled with `gcc -Wpedantic`
due to the placement of the casts in the initializer expressions, and
errors with Clang. In the worst case, these arrays could even cause C
compilers to perform unwanted type conversions.

To fix the issue, we now include special cases for `Array` and `Struct`
such that they are translated using `InitArray`, which does not contain
the type of the array and thus is pretty printed without explicit casts.
For example, this means that a nested array will be translated as `{{0,
1}, {2, 3}, {4, 5}}` instead of `{ (int32_t[2]){0, 1}, (int32_t[2]){2,
3}, (int32_t[2]){4, 5} }`. The same approach is used for structs, as the
  initialization syntax is compatible in C and language-c99 does not
provide specific construct for struct initializer, but that is an
artifact of how `language-c99-simple`'s API works.

Notes from Ivan Perez: work done by Ryan Scott; I have modified comments
and style.

Co-authored-by: Ivan Perez <ivan.perezdominguez@nasa.gov>