Remove test for copilot-core:Copilot.Core.Error. Refs #300.
The module copilot-core:Copilot.Core.Error only contains functions that
were deprecated in Copilot 3.6. Those functions should not be exposed,
since they serve no purpose outside of the internals of the
implementation.

As per our internal policy of waiting 3 versions from deprecation until
a public interface declaration can be removed, the complete module can
now be hidden.

Once the module is hidden, the tests that apply to that part of the
public interface of that module are no longer needed.

This commit removes those tests as a preliminary step for removing the
module itself. It also removes auxiliary operations that were only being
used by those tests.