Hide the module copilot-core:Copilot.Core.Error. Refs #300.
The module copilot-core:Copilot.Core.Error only contains functions that
were deprecated in Copilot 3.6. Those functions should not be exposed,
since they serve no purpose outside of the internals of the
implementation.

As per our internal policy of waiting 3 versions from deprecation until
a public interface declaration can be removed, the complete module can
now be hidden.

This commit modifies the cabal file so that the module is no longer
exposed, and modify Copilot.Core so that the module is also not exposed
indirectly (i.e., re-exported).