Move definitions from internal module back to copilot-core:Copilot.Core.Error. Refs #300.
The module copilot-core:Copilot.Core.Error only contains functions that
were deprecated in Copilot 3.6. Those functions should not be exposed,
since they serve no purpose outside of the internals of the
implementation.

As per our internal policy of waiting 3 versions from deprecation until
a public interface declaration can be removed. In order to follow that
policy, an intermediate, internal module containing the actual original
implementations was created and used in copilot-core, and the
functions published in Copilot.Core were annotated with warnings to
let external users know they were deprecated.

Since the module Copilot.Core.Error is now internal to copilot-core,
there is no need for having a separate internal module for the
implementations, and they can be returned to their original location.

This commit moves the definitions in Copilot.Core.ErrorInternal back to
Copilot.Core.Error, and adjusts imports throughout the library as
needed.