copilot-c99: Use type information to translate Recip. Refs #263.
The code generated for Recip (reciprocal fraction) currently does not
depend on the type of the arguments passed. That means that the C
sentence generated assumes that the type is double, although Recip also
supports Float. The code generated is the same, because it is not cast
explicitly and Recip imposes a requirement that the arguments be
Fractional, the code is semantically incorrect.

This commit modifies the translation of that Op1 into C99 so that the
type of the variables is taken into account.

It makes the function constNumTy top-level in the module so that it can
be used in the implementation of the translation for Recip.

Co-authored-by: Ryan Scott <rscott@galois.com>