copilot-c99: Use type information to translate abs. Refs #263.
At present, the C99 backend produces C code that translates the absolute
value function as abs, without taking into account the types of the
variables it is operating on. This has the undesired effect of producing
casts between integers, and single- and double-precision floating-point
numbers, which can lead to unexpected results.

This commit introduces a function that handles the translation of the
absolute value function (abs) by picking the right variant to use based
on the type of the operands.

Co-authored-by: Ryan Scott <rscott@galois.com>