copilot-c99: Use type information to translate functions from math.h. Refs #263.
At present, the C99 backend produces C code that executes mathematical
functions without taking into account the types of the variables it is
operating on. This has the undesired effect of producing casts between
single- and double-precision floating-point numbers, which can lead to
unexpected results.

This commit modifies the C code generation so that the types of
variables is taken into account in all translations involving
floating-point functions in math.h, when that C module provides
specialized variants for single- and double-precision floating-point
numbers.

Co-authored-by: Ryan Scott <rscott@galois.com>