copilot-c99: Adapt to work with language-c99-simple >= 0.2. Refs #371.
The new (0.2) version of the language-c99-simple library now requires
constructing some non-empty list with a dedicated value constructor
(NonEmpty) as opposed to using Haskell lists. This change was made to
better fit the C99 standard. This makes the current copilot-c99 code not
compile with the new version of the library.

This commit adapts copilot-c99 so that the new value constructor
NonEmpty is used where required.