copilot-c99: Update version bounds on language-c99*. Refs #371.
Copilot only works with language-c99 < 0.2, but that limits installation
with modern versions of GHC.

This commit adjusts the cabal file so that installation with modern
versions of language-c99, and consenquently GHC, is allowed. Note that
the API has changed, so further adaptations are required for code to
work with the new interface.