copilot-language: enable CLI args in tests. Refs #337.
This commit modifies the implementation of the test executable in
copilot-language so that the test seed can be passed as argument.

The change allows tests to be reproduced locally (e.g., with docker), by
feeding the same seed for random number generation to the test
executable, which is necessary to reproduce some errors.