copilot-language: Fix error in test case generation. Refs #337.
The tests generated for integral-carrying streams with bit functions
applied to them produce streams that cannot be evaluated. In particular,
when using one of the shifting operations, the second argument applied
to shift (the magnitude of the shift) may be negative, or be an integer
that becomes a negative number when cast to the necessary type, which
results in an exception.

This commit modifies the generation of bit-shifted streams by using
constant positive numbers for the magnitude of shifting.