copilot-core: Deprecate module Copilot.Core.External. Refs #322.
The module copilot-core:Copilot.Core.External is not being used by any
other part of Copilot. It is not a necessary part of the interface to
use Copilot, and not required by Copilot's requirements.

Following our policy of waiting 3 versions from deprecation until a
public interface declaration can be removed, this module is now
deprecated so that users who may need such features can report it,
and/or adapt their code as needed.