copilot-core: Reorganize module making definitions auxiliary. Refs #348.
As the type ShowWit is now completely hidden from the interface and only
used locally in this module, it has become auxiliary. This fact is best
reflected by moving the definitions to a location in the file where
their nature is made clear.

This commit moves the type definition, and the only function that can be
used to operate with it, to the bottom of the file, into an auxiliary
section. The decision to place them there (as opposed to e.g., in a
where clause) is made because the type cannot be defined in a where
clause.