copilot-core: Introduce internal module to deprecate Show. Refs #330.
The module Copilot.Core.Type.Show must be deprecated with a type
re-exported from a different module. The module itself will remain in
the implementation, but not be exposed to users.

In order to deprecate the module without triggering a warning while
compiling this library, this commit introduces a new module that
implements the content of Copilot.Core.Type.Show, and can be used during
the transition to compile and re-export entities without producing
warnings.