copilot-core: Adjust imports, options to avoid warnings. Refs #330.
The module Copilot.Core.Type.Show has been deprecated so that it can be
hidden from users. During this transition period, the module
Copilot.Core.Type.ShowInternal can be used instead.

This commit adjusts imports in the copilot-core package so that they use
Copilot.Core.Type.ShowInternal instead of the deprecated module. Where
that is not possible (i.e., in the tests), the commit uses compiler
options to disable warnings temporarily.