Configure travis to wait longer. Refs #335.
Some of the new versions of dependencies take too long to compile. For
example, what4 1.3 takes more than 10 minutes to compile with GHC 9, and
Travis CI cancels the build.

This commit prefixes the installation command with travis_wait to
indicate to Travis CI that it should wait even if not output is produced
for 10 minutes.