copilot-c99: Correct repo location, subdir in cabal file. Refs #333.
This commit modifies the location and the subdir of the
source-repository section of the cabal file. The location is adjusted to
use https, since unauthenticated git protocol is no longer supported by
git and authentication requires a user and key (which not all users may
have). The subdir is modified to match the new organization of the
copilot repo (all packages at top level without a lib subdirectory).