copilot-core: deprecate module Copilot.Core.PrettyDot. Refs #359.
The module Copilot.Core.PrettyDot is not being used by any other part of
Copilot or, as far as we know, by any user of Copilot. It is not a
necessary part of the interface to use Copilot, and not required by
Copilot's requirements. Although the idea is good, there is consensus
that it is not needed at this point.

Therefore, it would be suitable to remove this module.

This commit deprecates the module so that it can be removed in a future
version.