copilot-language: Deprecate prettyPrint. Refs #362.
The function copilot-language:Copilot.Language.prettyPrint is an aid to
reify a Spec and then pretty print the resulting Copilot Core
representation.

No part of Copilot needs this direct access. The module dependencies
that result from this change do not match those in our proposed design.
Consequently, it would be best to remove this function.

This commit deprecates the function prettyPrint to warn users that it
will be removed in the future.