copilot-core: Remove tests of module Copilot.Core.Type.Dynamic. Refs #360.
The module Copilot.Core.Type.Dynamic is not being used by any other part
of Copilot.

It was deprecated in Copilot 3.8 and, per our policy for removal of
outdated features, is now being removed.

This commit removes all tests pertaining to that module, as they will no
longer be necessary.