copilot-language: Reimplement DynStableName without unsafeCoerce. Refs #262.
The library copilot-language uses unsafeCoerce. This use of unsafeCoerce
can be avoided by changing the definition of
System.Mem.StableName.Dynamic.DynStableName and introducing an
existential type (akin to how the library data-reify defines
DynStableName).

In our effort to move Copilot to Class D, we have dictated that uses of
unsafe code (including functions specifically marked as unsafe in
Haskell) should be avoided unless strictly necessary.

This commit modifies DynStableName to be an existential type, which
eliminates the need for unsafeCoerce. The implementation of equality now
uses eqStableName. The import of unsafeCoerce is removed, and the module
is marked as Safe instead of just TrustWorthy.