copilot-c99: Fix error handling buffers in generated code for 'step'. Refs #314.
Appending values to a stream of arrays should delay the samples, but is
currently producing incorrect code.

The issue is related to how elements are stored in the intermediate
buffer associated to each stream, and how they are copied over. A part
of the generated C code, one can find copying operations like:

```C
(memcpy)((s0_tmp), ((s0_gen)()), (2));
(memcpy)(((s0)[s0_idx]), (s0_tmp), (2));
```

However, these operations do not take the size of the element into
account. For example, the code above was generated from a specification
for which, s0_gen returns an array with two int16_t elements.
Consequently, not all data is copied over.

This commit modifies the code generator so that memcpy operations use
sizeof to determine the amount of memory that must be copied.