copilot-core: Deprecate interpreter in copilot-core. Refs #361.
The architecture of Copilot mixes backends with other libraries, which makes
the design not uniform. For example, the interpreter is part of copilot-core,
while the C99 backend is in a separate library.

A prior commit addresses this issue by introducing a separate library
copilot-interpreter, and moving all interpretation definitions there. With
that new library, it makes sense to remove all such definitions from
copilot-core.

In conformance with our deprecation policy, this commit deprecates all public
definitions in copilot-core pertaining to interpretation of specifications.