copilot-interpreter: Add new library copilot-interpreter. Refs #361.
The architecture of the implementation of Copilot mixes backends with other
libraries, which makes the design not uniform. For example, the interpreter is
in copilot-core, while the C99 backend is in a separate library. The
implementation does not match our desired architecture according to our design
documents, and it makes the code partly harder to maintain.

This commit introduces a new library copilot-interpreter that provides an
interpreter backend for Copilot.

The imports from the Show module, which will no longer be exported in
copilot-core, are replicated as needed in the copilot-interpret package. A
flag disabled in a GHC pragma has been removed from the original module brought
over from copilot-core, as it is no longer removed.

The new library is enabled on Travis so that the tests are executed by the CI
server.