copilot-language: Use interpreter from copilot-interpreter. Refs #361.
The architecture of Copilot mixes backends with other libraries, which makes
the design not uniform. To address this issue, the interpreter in copilot-core
has been deprecated in favor of a new interpreter in copilot-interpreter.

This commit uses the modules in copilot-interpreter in the implementation of
copilot-language, as opposed to the ones from copilot-core.