copilot-core: Deprecate definitions in Array. Refs #369.
Multiple definitions in Copilot.Core.Type.Array are not being used by
any other part of Copilot or, as far as we know, by any user of Copilot.
These include the InnerType class, the Flatten class, the Foldable
instance, and the size function. These functions are now scheduled for
removal.

As per our internal policy of waiting 3 versions from deprecation until
a public interface declaration can be removed, the following are now
deprecated:

- Copilot.Core.Type.Array.InnerType
- Copilot.Core.Type.Array.Flatten
- Copilot.Core.Type.Array.size
- Instance of Data.Foldable for Arrays.

For the case of the instance, there does not seem to exist a deprecation
method that will make GHC warn users appropriately. For lack of a better
way, we document the deprecation in the instance haddock documentation.