copilot-core: Remove unnecessary type constraints. Refs #369.
Multiple definitions in Copilot.Core.Type.Array are not being used by
any other part of Copilot or, as far as we know, by any user of Copilot.
These include the InnerType class, the Flatten class, the Foldable
instance, and the size function.

The classes InnerType and Flatten are being used as type constraints in
other functions, but they are not really required: removing them has no
effect in the code.

This commit removes mentions of the aforementioned classes from modules
in copilot-core where they are no longer needed.