copilot-c99: Require language-c99-simple-0.2.2. Refs #314.
Appending values to a stream of arrays should delay the samples, but is
currently producing incorrect code. To address that issue, we need to
generate memcpy operations that take the size of the elements being
copied into account, for which we need to use a special construct to
produce a sizeof in C99 that is only available in version 0.2.2 of
language-c99-simple.

To make that solution possible, this commit modifies the cabal file so
that copilot-c99 requires that specific version of language-c99-simple
(or greater).